+++ js/src/packs/bqplot.js
@@ -1,0 +1,6 @@
+import * as bqplot from "bqplot";
+import * as bqscales from "bqscales";
+import "bqplot/css/bqplot.css";
+
+window.__tkipwRegisterPack("bqplot", bqplot);
+window.__tkipwRegisterPack("bqscales", bqscales);