+++ js/src/packs/ipycanvas.js
@@ -1,0 +1,3 @@
+import * as ipycanvas from "ipycanvas";
+
+window.__tkipwRegisterPack("ipycanvas", ipycanvas);