+++ js/src/packs/leaflet.js
@@ -1,0 +1,4 @@
+import * as jupyterLeaflet from "jupyter-leaflet";
+import "leaflet/dist/leaflet.css";
+
+window.__tkipwRegisterPack("jupyter-leaflet", jupyterLeaflet);